server/requirements.txt
fastapi
tortoise-orm
starlette
python-dotenv
uvicorn
pillow
pyjwt
sqlite3
pydantic
fastapi-mail